+++ aws-sdk-core/lib/aws-sdk-core.rb
@@ -97,6 +97,11 @@ module Aws
     MarketplaceCommerceAnalytics
     MarketplaceEntitlementService
     MarketplaceMetering
+    MediaConvert
+    MediaLive
+    MediaPackage
+    MediaStore
+    MediaStoreData
     MigrationHub
     Mobile
     OpsWorks